+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ### Pulse
 
-- Address [#162](https://github.com/kean/Pulse/issues/162): `LoggerStore/storeRequest` now accepts a custom label
+- Address [#162](https://github.com/kean/Pulse/issues/162): Add `label` parameter to `LoggerStore/storeRequest` and `NetworkLogger/Configuration` to allow customizing the label associated with the created logs 
 - Remove `NetworkLoggerInsights`
 
 ## Pulse 3.4.3